Refactor: remove unnecessary entities and changes data type from int to double

+++ src/com/tw/bootcamp/p1/entities/Shape.java
@@ -1,7 +1,7 @@
 package com.tw.bootcamp.p1.entities;
 
 public interface Shape {
-    int area();
+    double area();
 
-    int perimeter();
+    double perimeter();
 }
